+++ dockermap/__init__.py
@@ -2,7 +2,7 @@
 from __future__ import unicode_literals
 
 
-__version__ = '0.6.5'
+__version__ = '0.6.6b1'
 
-DEFAULT_BASEIMAGE = 'tianon/true'
-DEFAULT_COREIMAGE = 'busybox'
+DEFAULT_BASEIMAGE = 'tianon/true:latest'
+DEFAULT_COREIMAGE = 'busybox:latest'