+++ dockermap/map/policy/cache.py
@@ -26,7 +26,7 @@ class CachedImages(CachedItems, dict):
     Dictionary of image names and ids, which also keeps track of the client object to pull images if necessary.
     """
     def __init__(self, *args, **kwargs):
-        self._latest = set()
+        self._updated = set()
         super(CachedImages, self).__init__(*args, **kwargs)
 
     def refresh(self):
@@ -42,21 +42,23 @@ class CachedImages(CachedItems, dict):
             if tags:
                 self.update({tag: image['Id'] for tag in tags})
 
-    def reset_latest(self):
+    def reset_updated(self):
         """
-        Resets the cache which images have been pulled (i.e. updated to the latest version.)
+        Resets the cache which images have been pulled (i.e. updated on the default tag.)
         """
-        self._latest = set()
+        self._updated = set()
 
-    def ensure_image(self, image_name, pull_latest=False, insecure_registry=False):
+    def ensure_image(self, image_name, pull=False, insecure_registry=False):
         """
         Ensures that a particular image is present on the client. If it is not, a new copy is pulled from the server.
 
         :param image_name: Image name. If it does not include a specific tag, ``latest`` is assumed.
         :type image_name: unicode | str
-        :param pull_latest: If the image includes a latest-tag, pull it from the server. This is is done only once in
-         for the lifecycle of the cache, or unless `:meth:reset_latest` is called.
-        :type pull_latest: bool
+        :param pull: If the image includes a tag, pull it from the server even if it exists. This is is done only once
+          in for the lifecycle of the cache, or unless `:meth:reset_updated` is called.
+        :type pull: bool
+        :param insecure_registry: Pull from an insecure registry where necessary.
+        :type insecure_registry: bool
         :return: Image id associated with the image name.
         :rtype: unicode | str
         """
@@ -67,14 +69,13 @@ class CachedImages(CachedItems, dict):
             full_name = '{0}:latest'.format(image_name)
             image = image_name
             tag = 'latest'
-        update_latest = pull_latest and tag == 'latest' and full_name not in self._latest
-        if update_latest or full_name not in self:
+        if (pull and full_name not in self._updated) or full_name not in self:
             self._client.pull(repository=image, tag=tag, insecure_registry=insecure_registry)
             images = self._client.images(name=image_name)
             for new_image in images:
                 tags = new_image.get('RepoTags')
                 if tags:
-                    self._latest.update(tags)
+                    self._updated.update(tags)
                     self.update({tag: new_image['Id'] for tag in tags})
         try:
             return self[full_name]